restaurant 관련 수정

+++ routes/restaurant.js
@@ -2,16 +2,38 @@ var express = require('express');
 var router = express.Router();
 var connection = require("../lib/db");
 
-router.get('/add', function (req, res, next) {
-  res.render('addRestaurant', { title: 'Add New Restaurant', action: 'add', name: '', address: '' });
+router.get('/add', async (req, res, next) => {
+  try{
+    const [tags] = await connection.execute('SELECT id, name FROM tags');
+    res.render('addRestaurant', { title: 'Add New Restaurant', action: 'add', name: '', address: '', tags: tags});
+  }
+  catch (err){
+    next(err);
+  }
 });
 
 
 router.post('/add', async (req, res, next) => {
   try{
-    const [rows] = await connection.query('INSERT INTO restaurants (name, address) VALUES (?,?)', 
-                      [req.body.name, req.body.address]);
-    res.redirect('/');
+    if (req.body.name == null || req.body.name.length == 0) {
+      const [tags] = await connection.execute('SELECT id, name FROM tags');
+      res.render('addRestaurant', { title: 'Add New Restaurant', action: 'add', name: '', address: req.body.address, tags: tags, error: "Please Insert Restaurant's Name!"});
+    }
+    else {
+      const [rows] = await connection.query('INSERT INTO restaurants (name, address) VALUES (?,?)', 
+                        [req.body.name, req.body.address]);
+
+      if(req.body.tags != null){
+        let tags = null
+        if(Array.isArray(req.body.tags))
+          tags = req.body.tags.map(x=>[rows.insertId, x])
+        else
+          tags = [[rows.insertId, req.body.tags]]
+
+        await connection.query('INSERT INTO mapping_tag_restaurant (restaurant_id, tag_id) VALUES ?', [tags]);
+      }
+      res.redirect('/');
+    }
   }
   catch(err){
     next(err)
@@ -70,17 +92,28 @@ router.get('/:id/edit', async (req, res, next) => {
 
 router.post('/:id/edit', async (req, res, next) => {
   try{
-    await connection.query('UPDATE restaurants SET name = ?, address = ? WHERE id = ?', 
-                      [req.body.name, req.body.address, req.params.id]);
-    await connection.query('DELETE FROM mapping_tag_restaurant WHERE restaurant_id = ?', [req.params.id]);
-    let tags = null
-    if(Array.isArray(req.body.tags))
-      tags = req.body.tags.map(x=>[req.params.id, x])
-    else
-      tags = [[req.params.id, req.body.tags]]
-    console.log(tags)
-    await connection.query('INSERT INTO mapping_tag_restaurant (restaurant_id, tag_id) VALUES ?', [tags]);
-    res.redirect('/restaurant/'+req.params.id);
+    if (req.body.name == null || req.body.name.length == 0) {
+      const [tags] = await connection.query("SELECT tags.id, tags.name, m.restaurant_id FROM tags LEFT OUTER JOIN mapping_tag_restaurant AS m "
+      + "ON tags.id = m.tag_id "
+      + "AND m.restaurant_id = ?", [req.params.id])
+      res.render('addRestaurant', { title: 'Edit Restaurant', action: req.params.id + "/edit", name: '', address: req.body.address, tags: tags, error: "Please Insert Restaurant's Name!"});
+    }
+    else {
+      await connection.query('UPDATE restaurants SET name = ?, address = ? WHERE id = ?', 
+                        [req.body.name, req.body.address, req.params.id]);
+      await connection.query('DELETE FROM mapping_tag_restaurant WHERE restaurant_id = ?', [req.params.id]);
+      
+      if(req.body.tags != null){
+        let tags = null
+        if(Array.isArray(req.body.tags))
+          tags = req.body.tags.map(x=>[req.params.id, x])
+        else
+          tags = [[req.params.id, req.body.tags]]
+        console.log(tags)
+        await connection.query('INSERT INTO mapping_tag_restaurant (restaurant_id, tag_id) VALUES ?', [tags]);
+      }
+      res.redirect('/restaurant/'+req.params.id);
+    }
   }
   catch(err){
     next(err)
